requirements.txt
Django>=4.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.5.0
mysqlclient==2.2.7
django-model-utils
gunicorn==20.1.0
pytest
pytest-django
factory-boy==3.3.0
